+++ src/main/java/model/sudokuFramework_Candidates/exceptions/WrongValueException.java
@@ -1,4 +1,4 @@
-package model.sudokuWithCandidatesQuickAndDirty.exceptions;
+package model.sudokuFramework_Candidates.exceptions;
 
 public class WrongValueException extends Exception{
