+++ src/main/java/model/sudokuFramework_Candidates/interfaces/FieldTypes.java
@@ -1,0 +1,5 @@
+package model.sudokuFramework_Candidates.interfaces;
+
+public enum FieldTypes {
+    SudokuField,
+}
