+++ src/main/java/model/sudokuFramework_Candidates/interfaces/Solver.java
@@ -1,0 +1,5 @@
+package model.sudokuFramework_Candidates.interfaces;
+
+public interface Solver {
+    public boolean solve();
+}
