--- src/main/java/model/sudokuWithCandidatesQuickAndDirty/interfaces/FieldFactoryInterface.java
@@ -1,7 +1,0 @@
-package model.sudokuWithCandidatesQuickAndDirty.interfaces;
-
-import model.sudokuWithCandidatesQuickAndDirty.fields.Field;
-
-public interface FieldFactoryInterface {
-    public Field getInstance(FieldTypes type, int value);
-}
