--- src/main/java/model/sudokuWithCandidatesQuickAndDirty/interfaces/FieldTypes.java
@@ -1,5 +1,0 @@
-package model.sudokuWithCandidatesQuickAndDirty.interfaces;
-
-public enum FieldTypes {
-    SudokuField,
-}
