--- src/main/java/model/sudokuWithCandidatesQuickAndDirty/interfaces/Solver.java
@@ -1,5 +1,0 @@
-package model.sudokuWithCandidatesQuickAndDirty.interfaces;
-
-public interface Solver {
-    public boolean solve();
-}
