--- src/main/java/model/sudokuWithCandidatesQuickAndDirty/interfaces/SolverFactoryInterface.java
@@ -1,7 +1,0 @@
-package model.sudokuWithCandidatesQuickAndDirty.interfaces;
-
-import model.sudokuWithCandidatesQuickAndDirty.boards.AbstractBoard;
-
-public interface SolverFactoryInterface {
-    public Solver getInstance(FieldTypes type, AbstractBoard board);
-}
